+++ src/util/index.ts
@@ -1,4 +1,4 @@
 export * from "./dom";
 export * from "./misc";
 export * from "./tags";
-export * from "./heroActions"
+export * from "./hero-actions"
